src/third_party/blink/web_tests/fast/html/imports/import-expando-gc-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
Properties in imported dom nodes should be retained.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getNodeInImport().foo is "Hello"
PASS getNodeInImport().foo is "Hello"
PASS successfullyParsed is true

TEST COMPLETE
